requirements.txt
biopython==1.85
certifi==2025.8.3
charset-normalizer==3.4.3
ete3==3.1.3
idna==3.10
joblib==1.5.2
legacy-cgi==2.6.3
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.2
legacy-cgi
# orthofinder @ git+https://github.com/OrthoFinder/OrthoFinder.git@1851568b340c4459ab1662516a364c00515d4901
Pygments==2.19.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.1
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
urllib3==2.5.0
pandas
torch
matplotlib
